+++ CustomException.cs
@@ -3,6 +3,9 @@ using System.Net;
 
 namespace Utility.BaseClasses
 {
+    /// <summary>
+    /// Extends the Exception class to give you a HttpStatusCode and Message as well as the base Exception information.
+    /// </summary>
     public class CustomException : Exception
     {
         public HttpStatusCode Status { get; set; }