+++ Release Notes.md
@@ -4,6 +4,11 @@ The Release Notes for Utility.BaseClasses contains information about when change
 
 ***
 
+**Update 1.0.2 - 01/10/2024**
+* Added a new class HttpRequest.UserAgent for use in web service requests.
+* Added .NET 8 to the TargetFrameworks list.
+* Updated copyright to 2024.
+
 **Update 1.0.1 - 03/06/2023**
 * Corrected an issue with a CS0122 error on the Error class.